+++ src/parameterestimation.jl
@@ -1,3 +1,4 @@
 include(joinpath("parameterestimation", "bayesian.jl"))
 include(joinpath("parameterestimation", "maximumlikelihood.jl"))
 include(joinpath("parameterestimation", "probabilityweightedmoment.jl"))
+